requirements.txt
aniso8601==1.2.0
asn1crypto==0.22.0
boto3==1.4.6
botocore==1.6.3
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
colorclass==2.2.0
cryptography==2.0.3
docutils==0.14
Flask==0.12.1
Flask-Ask==0.9.7
formats==0.1.1
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.9.3
MarkupSafe==1.0
pycparser==2.18
pyOpenSSL==17.0.0
python-dateutil==2.6.1
PyYAML==3.12
Quizlet-API==0.1
requests==2.18.4
s3transfer==0.1.10
six==1.10.0
tortilla==0.4.2
urllib3==1.22
Werkzeug==0.12.2
